[Docs] Update API docs (#27)
* Fix circular dependencies

* Remove unused imports

* Sorted imports
